+++ src/App.tsx
@@ -1,24 +1,31 @@
 import React, { useState } from 'react';
+
 import './App.css';
+import { getNumbers } from './utils';
 import { Pagination } from './components/Pagination';
+
+const items = getNumbers(1, 42)
+  .map((n: number) => `Item ${n}`);
+
+function getCurrentItems(currentPage:number, perPage: number): string[] {
+  if (currentPage === 1) {
+    return [...items].splice(0, perPage);
+  }
+
+  return [...items].splice((currentPage - 1) * perPage, perPage);
+}
 
 export const App: React.FC = () => {
   const [perPage, setPerPage] = useState('3');
   const [currentPage, setCurrentPage] = useState(1);
-  const [fromPage, setFromPage] = useState('1');
-  const [toPage, setToPage] = useState('3');
-
-  function getFromPage(page: number) {
-    setFromPage(String((page - 1) * +perPage + 1));
-    setToPage(String((page - 1) * +perPage + +perPage));
-  }
+  const currentItems: string[] = getCurrentItems(+currentPage, +perPage);
 
   return (
     <div className="container">
       <h1>Items with Pagination</h1>
 
       <p className="lead" data-cy="info">
-        {`Page ${currentPage} (items ${fromPage} - ${toPage} of 42)`}
+        {`Page ${currentPage} (items ${parseInt(currentItems[0], 10)} - ${parseInt(currentItems[currentItems.length - 1], 10)} of 42)`}
       </p>
 
       <div className="form-group row">
@@ -27,7 +34,10 @@ export const App: React.FC = () => {
             data-cy="perPageSelector"
             id="perPageSelector"
             className="form-control"
-            onChange={(event) => setPerPage(event.target.value)}
+            onChange={(event) => {
+              setPerPage(event.target.value);
+              setCurrentPage(1);
+            }}
           >
             <option value="3">3</option>
             <option value="5">5</option>
@@ -47,9 +57,14 @@ export const App: React.FC = () => {
         currentPage={currentPage}
         onPageChange={(page) => {
           setCurrentPage(page);
-          getFromPage(page);
         }}
       />
+
+      <ul>
+        {[...currentItems].map(el => (
+          <li data-cy="item" key={el}>{el}</li>
+        ))}
+      </ul>
     </div>
   );
 };
